Remove singleton SystemDisplay
